Add Lower Treshold to Deregister Tracked Objects and Removed Count Marker Lag
Add implementation of the lower treshold that removes objects from the list of tracked objects once they are below a certain
treshold in the frame. Also, to remove the coun tracker lag that was previously present now when objec is lost the count
is no longer printed on the frame.